+++ main.py
@@ -9,11 +9,14 @@ import uuid
 from pathlib import Path
 import io
 import codecs
-
+from appNumberGenerator import appNoGenerator
+from randomStringGenerator import randomStringGen
+from serialNumberGenerator import serialNoGenerator
+from numberGenerator import numberGenerator
 # Пути к файлам JSON
 path = 'C:/Users/vtrofimenko/Desktop/Test-knife/logs.txt'
 jsonFPAPath = 'C:\\Users\\vtrofimenko\\Desktop\\tnife\\Tnife-main\\fpaProsses\\FPA.json'
-jsonASRPATH = 'C:\\Users\\vtrofimenko\\Desktop\\tnife\\Tnife-main\\arfProsses\\ARF.json'
+jsonASRPATH = 'C:\\Users\\vtrofimenko\\Desktop\\Test-knife\\arfProsses\\ARF.json'
 
 message = ''' 
 -case -count - для генерации caseId для создания дел, при использовании нужно указывать количество генераций 
@@ -27,67 +30,11 @@ message = '''
 Args = sys.argv
 
 
-def appNoGenerator():
-    length_start = 1000000000000
-    length_end = 9000000000000
-    return str(random.randrange(length_start, length_end))
-
-
-def printForJson():
-    stringUUid = []
-    for _ in range(2):
-        myUuid = uuid.uuid4()
-        stringUUid.append(str(myUuid))
-    result = appNoGenerator() + ", " + ', '.join(stringUUid)
-    return result
-
-# переделать функции на возврат одного значения 
-def uuidsCountPrinter(secondArg):
-    stringUUid = []
-    for _ in range(secondArg):
-        myUuid = uuid.uuid4()
-        stringUUid.append(str(myUuid))
-        result = '\n'.join(stringUUid)
-    resultOut = ','.join(stringUUid)
-    result = '\n'.join(stringUUid)
-    print(str(result))
-    return resultOut
-
-
-def appNoPrinter(secondArg):
-    result = '\n'.join([appNoGenerator() for _ in range(secondArg)])
-    pyperclip.copy(result)
-    return result.replace('\n', ', ')
-
-
-def copyCoreId(secondArg):
-    result = '\n'.join([generate_random_string() for _ in range(secondArg)])
-    pyperclip.copy(result)
-    return result.replace('\n', ', ')
-
-
-def generate_random_string():
-    start_length = 1000000000000000000
-    end_length = 8999999999999999999
-    return str(random.randrange(start_length, end_length))
-
-
 def write_log(text):
     current_time = strftime("%Y-%m-%d %H:%M:%S")
     with open(path, 'a') as f:
         f.write(current_time + " >>>> " + str(text) + '\n')
 
-
-def serGenerator():
-    start = 1000
-    end = 9999
-    return (str(random.randrange(start, end)))
-
-
-def numberGenerator():
-    start = 1000000
-    end = 9999999
-    return (str(random.randrange(start, end)))
 
 def epgu(secondArg):
     stringUUid = [str(uuid.uuid4()) for _ in range(2)]
@@ -112,7 +59,7 @@ def epgu(secondArg):
         root = ET.fromstring(xml_data)
 
         for orderId in root.iter('{urn://mvd/gismu/arf/1.1.0}orderId'):
-            orderId.text = appNoGenerator()
+            orderId.text = appNoGenerator.appNoGenerator()
             orderId_text = orderId.text
             break
         if orderId_text is None:
@@ -145,7 +92,7 @@ def epgu(secondArg):
         namespace1 = {'ns1': 'urn://mvd/gismu/FPA/EPGU/types/1.0.6'}
 
         for orderId in root.findall('.//ns0:OrderId', namespace0):
-            orderId.text = appNoGenerator()
+            orderId.text = appNoGenerator.appNoGenerator()
             orderId_text = orderId.text
             break
 
@@ -153,12 +100,12 @@ def epgu(secondArg):
             orderId_text = ""
 
         for series in root.findall('.//ns1:Series', namespace1):
-            series.text = serGenerator()
+            series.text = serialNoGenerator.serGenerator()
             series_text = series.text
             break
 
         for number in root.findall('.//ns1:Number', namespace1):
-            number.text = numberGenerator()
+            number.text = numberGenerator.numberGenerator()
             number_text = number.text
             break
 
@@ -170,19 +117,26 @@ def epgu(secondArg):
             json.dump(data, f, indent=4, ensure_ascii=False)
 
         return f"messageId = {messageId}, clientId = {clientId}, OrderId = {orderId_text} , Series = {series_text} , Number = {number_text}"
+    
+    
+from printForJson import printForJson
+from coreCaseIdGenerator import coreCaseIdGenerator
+from appNumberPrinter import appNumberPrinter
+from uuidGenerator import uuidGenerator
+
 
 if __name__ == '__main__':
     if len(Args) > 1:
         if Args[1] == "-json":
-            write_log(printForJson())
+            write_log(printForJson.printForJson())
         elif Args[1] == "-case":
             if len(Args) > 2:
-                write_log(copyCoreId(int(Args[2])))
+                write_log(coreCaseIdGenerator.copyCoreId(int(Args[2])))
             else:
                 print("Недостаточно аргументов для -case.")
         elif Args[1] == "-app":
             if len(Args) > 2:
-                write_log(appNoPrinter(int(Args[2])))
+                write_log(appNumberPrinter.appNoPrinter(int(Args[2])))
             else:
                 print("Недостаточно аргументов для -app.")
         elif Args[1] == "-epgu":
@@ -192,7 +146,7 @@ if __name__ == '__main__':
                 print("Недостаточно аргументов для -epgu.")
         elif Args[1] == "-uuid":
             if len(Args) > 2:
-                write_log(uuidsCountPrinter(int(Args[2])))
+                write_log(uuidGenerator.uuidsCountPrinter(int(Args[2])))
             else:
                 print("Недостаточно аргументов для -uuid.")
         else:
