+++ numberGenerator.py
@@ -1,0 +1,7 @@
+import random
+class numberGenerator:
+    @staticmethod
+    def numberGenerator():
+        start = 1000000
+        end = 9999999
+        return (str(random.randrange(start, end)))
