+++ serialNumberGenerator.py
@@ -1,0 +1,8 @@
+import random
+
+class serialNoGenerator:
+    @staticmethod
+    def serGenerator():
+        start = 1000
+        end = 9999
+        return (str(random.randrange(start, end)))
